+++ src/Controller/HomeController.php
@@ -8,20 +8,17 @@ use App\Controller\TemplateController;
 use App\Service\HomeService;
 use App\Service\UserService;
 
-class HomeController
+class HomeController extends BaseController
 {
-    private $prefix = '';
     private static $instance;
     private HomeService $homeService;
     private UserService $userService;
 
-    private function __construct(string $target)
+    private function __construct(string $ajaxMode)
     {
-        if ($target) {
-            $this->prefix = 'alone_';
-        }
         $this->homeService = HomeService::getInstance();
         $this->userService = UserService::getInstance();
+        parent::__construct($ajaxMode);
     }
 
     public static function getInstance(string $target): self
@@ -35,15 +32,12 @@ class HomeController
     public function displayHome(int $id = 1): void
     {
         $profile = $this->homeService->getHome($id);
-        $template_page = $this->prefix . 'profile';
-        $template = new TemplateController($template_page);
         $array['results'] = $profile;
         $links = $this->userService->getLinks($id);
         foreach ($links as $link) {
             $array['links'][] = ['url' => $link->url];
         }
-
-        $template->call($array);
+        $this->renderHtml($array, 'profile');
     }
 
 }