+++ src/Rooter.php
@@ -14,16 +14,17 @@ class Rooter
         //boot
     }
 
-    public function index(array $params): void
+    public function index(array $gets): void
     {
-        //pr($params);
-        $com = $params['com'] ?? 'test'; //default
-        $id = $params['p1'] ?? 1;
+        //pr($gets);
+        $com = $gets['com'] ?? 'home'; //default
+        $id = $gets['p1'] ?? 1; //default
         $target = get('_tg');
-        $articleController = ArticleController::getInstance($target);
-        $categoryController = CategoryController::getInstance($target);
-        $userController = UserController::getInstance($target);
-        $homeController = HomeController::getInstance($target);
+        $ajaxMode = $target ? 'true' : 'false';
+        $articleController = ArticleController::getInstance($ajaxMode);
+        $categoryController = CategoryController::getInstance($ajaxMode);
+        $userController = UserController::getInstance($ajaxMode);
+        $homeController = HomeController::getInstance($ajaxMode);
         match ($com) {
             'home' => $homeController->displayHome(1),
             'post' => $articleController->displayPost((int) $id),
@@ -33,6 +34,11 @@ class Rooter
             'categories' => $categoryController->displayCategories(),
             'user' => $userController->displayName($id),
             'profile' => $userController->displayProfile($id),
+            'register' => $userController->displayRegisterForm(),
+            'registerUser' => $userController->registerUser($gets),
+            'login' => $userController->loginRoot(),
+            'logon' => $userController->authentification($gets),
+            'logout' => $userController->logout($gets),
             default => $articleController->displayPost(1)
         };
     }