+++ src/Service/UserService.php
@@ -6,6 +6,8 @@ namespace App\Service;
 
 use App\Repository\userRepository;
 use App\Entity\UserEntity;
+use App\Model\UserModel;
+use App\Model\ErrorModel;
 
 class UserService
 {
@@ -33,6 +35,34 @@ class UserService
     public function getUser(int $id): UserEntity
     {
         return $this->userRepository->userInfos($id);
+    }
+
+    public function getUserFromName(string $name): UserEntity|bool
+    {
+        return $this->userRepository->findUserFromName($name);
+    }
+
+    public function getUserLoged(string $name, string $pswd): UserModel|ErrorModel
+    {
+        $userEntity = $this->userRepository->findUserFromName($name);
+
+        if (empty($userEntity->name)) {
+            return new ErrorModel(message: 'Utilisateur inconnu');
+        }
+
+        $isGoodPassword = password_verify($pswd, $userEntity->pswd ?? '');
+        if (!$isGoodPassword) {
+            return new ErrorModel(message: 'Mot de passe non reconnu');
+        }
+
+        return new UserModel(name: $userEntity->name, uid: $userEntity->id);
+    }
+
+    public function registerUser(string $name, string $mail, string $pswd): string
+    {
+        $pswd = password_hash($pswd, PASSWORD_DEFAULT);
+        $values = ['name' => $name, 'mail' => $mail, 'pswd' => $pswd];
+        return $this->userRepository->registerUser($values);
     }
 
     public function getLinks(int $id): array